requirements.txt
Flask==2.0.2
numpy==1.22.1
Pillow==9.0.0
